moved inline SVG's to separate files and optimized them.
